move eventMap to event package
